refactor: usize to u64

+++ crates/common/consensus/lean/src/block.rs
@@ -22,8 +22,8 @@ pub struct SignedBlock {
 // TODO: Add back #[derive(TreeHash)]
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize, Encode, Decode, Default)]
 pub struct Block {
-    pub slot: usize,
-    pub proposer_index: usize,
+    pub slot: u64,
+    pub proposer_index: u64,
     pub parent: Option<B256>,
     pub votes: VariableList<Vote, U16777216>,
     pub state_root: Option<B256>,
